// src/frontend/app/src/api-client/models/StreakState.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenBook API
 * Beautiful and Engaging Learning Materials
 *
 * The version of the OpenAPI document: 0.0.1-pre1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Current daily streak state of an account
 * @export
 * @interface StreakState
 */
export interface StreakState {
    /**
     * 
     * @type {number}
     * @memberof StreakState
     */
    currentStreak: number;
    /**
     * 
     * @type {number}
     * @memberof StreakState
     */
    longestStreak: number;
    /**
     * 
     * @type {Date}
     * @memberof StreakState
     */
    lastActiveDate: Date | null;
    /**
     * 
     * @type {number}
     * @memberof StreakState
     */
    streakFreezes: number;
}

/**
 * Check if a given object implements the StreakState interface.
 */
export function instanceOfStreakState(value: object): value is StreakState {
    if (!('currentStreak' in value) || value['currentStreak'] === undefined) return false;
    if (!('longestStreak' in value) || value['longestStreak'] === undefined) return false;
    if (!('lastActiveDate' in value) || value['lastActiveDate'] === undefined) return false;
    if (!('streakFreezes' in value) || value['streakFreezes'] === undefined) return false;
    return true;
}

export function StreakStateFromJSON(json: any): StreakState {
    return StreakStateFromJSONTyped(json, false);
}

export function StreakStateFromJSONTyped(json: any, ignoreDiscriminator: boolean): StreakState {
    if (json == null) {
        return json;
    }
    return {
        
        'currentStreak': json['current_streak'],
        'longestStreak': json['longest_streak'],
        'lastActiveDate': (json['last_active_date'] == null ? null : new Date(json['last_active_date'])),
        'streakFreezes': json['streak_freezes'],
    };
}

export function StreakStateToJSON(json: any): StreakState {
    return StreakStateToJSONTyped(json, false);
}

export function StreakStateToJSONTyped(value?: StreakState | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'current_streak': value['currentStreak'],
        'longest_streak': value['longestStreak'],
        'last_active_date': (value['lastActiveDate'] == null ? null : (value['lastActiveDate'] as any).toISOString().substring(0,10)),
        'streak_freezes': value['streakFreezes'],
    };
}