// src/frontend/app/src/api-client/models/CourseProgress.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenBook API
 * Beautiful and Engaging Learning Materials
 *
 * The version of the OpenAPI document: 0.0.1-pre1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Course Progress
 * @export
 * @interface CourseProgress
 */
export interface CourseProgress {
    /**
     * 
     * @type {string}
     * @memberof CourseProgress
     */
    readonly id: string;
    /**
     * 
     * @type {string}
     * @memberof CourseProgress
     */
    account: string;
    /**
     * 
     * @type {string}
     * @memberof CourseProgress
     */
    course: string;
    /**
     * 
     * @type {number}
     * @memberof CourseProgress
     */
    coursePoints?: number;
    /**
     * 
     * @type {number}
     * @memberof CourseProgress
     */
    courseLevel?: number;
    /**
     * 
     * @type {string}
     * @memberof CourseProgress
     */
    courseProgress?: string;
}

/**
 * Check if a given object implements the CourseProgress interface.
 */
export function instanceOfCourseProgress(value: object): value is CourseProgress {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('account' in value) || value['account'] === undefined) return false;
    if (!('course' in value) || value['course'] === undefined) return false;
    return true;
}

export function CourseProgressFromJSON(json: any): CourseProgress {
    return CourseProgressFromJSONTyped(json, false);
}

export function CourseProgressFromJSONTyped(json: any, ignoreDiscriminator: boolean): CourseProgress {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'account': json['account'],
        'course': json['course'],
        'coursePoints': json['course_points'] == null ? undefined : json['course_points'],
        'courseLevel': json['course_level'] == null ? undefined : json['course_level'],
        'courseProgress': json['course_progress'] == null ? undefined : json['course_progress'],
    };
}

export function CourseProgressToJSON(json: any): CourseProgress {
    return CourseProgressToJSONTyped(json, false);
}

export function CourseProgressToJSONTyped(value?: Omit<CourseProgress, 'id'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'account': value['account'],
        'course': value['course'],
        'course_points': value['coursePoints'],
        'course_level': value['courseLevel'],
        'course_progress': value['courseProgress'],
    };
}