// src/frontend/app/src/api-client/models/SkillProgress.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenBook API
 * Beautiful and Engaging Learning Materials
 *
 * The version of the OpenAPI document: 0.0.1-pre1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Skill Progress
 * @export
 * @interface SkillProgress
 */
export interface SkillProgress {
    /**
     * 
     * @type {string}
     * @memberof SkillProgress
     */
    readonly id: string;
    /**
     * 
     * @type {string}
     * @memberof SkillProgress
     */
    account: string;
    /**
     * 
     * @type {string}
     * @memberof SkillProgress
     */
    skill: string;
    /**
     * 
     * @type {number}
     * @memberof SkillProgress
     */
    level?: number;
    /**
     * 
     * @type {string}
     * @memberof SkillProgress
     */
    progress?: string;
}

/**
 * Check if a given object implements the SkillProgress interface.
 */
export function instanceOfSkillProgress(value: object): value is SkillProgress {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('account' in value) || value['account'] === undefined) return false;
    if (!('skill' in value) || value['skill'] === undefined) return false;
    return true;
}

export function SkillProgressFromJSON(json: any): SkillProgress {
    return SkillProgressFromJSONTyped(json, false);
}

export function SkillProgressFromJSONTyped(json: any, ignoreDiscriminator: boolean): SkillProgress {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'account': json['account'],
        'skill': json['skill'],
        'level': json['level'] == null ? undefined : json['level'],
        'progress': json['progress'] == null ? undefined : json['progress'],
    };
}

export function SkillProgressToJSON(json: any): SkillProgress {
    return SkillProgressToJSONTyped(json, false);
}

export function SkillProgressToJSONTyped(value?: Omit<SkillProgress, 'id'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'account': value['account'],
        'skill': value['skill'],
        'level': value['level'],
        'progress': value['progress'],
    };
}